# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GoShorty URL Shortener API",
    "version": "3.0.0",
    "description": "REST API for the GoShorty URL shortener (TTL links, custom codes, visits, stats).\n\nThe canonical API surface is `/api/v1/*`. Legacy `/api/*` routes remain fully functional as backward-compatible aliases and are marked `deprecated` in this document. The canonical short-link redirect is `/r/:code`; `/s/:code` and `/goshorty/:timeout/:code` are retained so previously issued links keep working.\n\nAll protected endpoints require `Authorization: Bearer <token>`. Tokens are HMAC-signed payload.signature strings issued by `/api/v1/auth/login` or `/api/v1/auth/register` (see SECURITY.md).",
    "contact": {
      "name": "Kouskous56",
      "url": "https://github.com/Kouskous56/Goshorty"
    }
  },
  "servers": [
    {
      "url": "http://goshorty.localhost:8080",
      "description": "Local default (configured via PUBLIC_BASE_URL)"
    }
  ],
...
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "description": "Creates a user and returns an access token in the same call.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created; token issued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request body or weak password",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Username already registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and obtain a bearer token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful; token issued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Current user information",
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeResponse" }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or invalid bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "User no longer exists",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password": {
      "put": {
        "tags": ["Auth"],
        "summary": "Change the current user's password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid body, weak or unchanged password",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Missing/invalid token or wrong current password",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/revoke": {
      "post": {
        "tags": ["Auth"],
        "summary": "Revoke all sessions for the current user",
        "description": "Invalidates the token used for the request and every other token issued to that user. Subsequent requests with old tokens return 401.",
        "responses": {
          "200": {
            "description": "All sessions revoked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/urls": {
      "post": {
        "tags": ["URLs"],
        "summary": "Create a short URL",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ShortenRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Short URL created; short_url uses the canonical /r/:code format",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShortenResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid request body or URL",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "Custom short code already taken",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["URLs"],
        "summary": "List URLs visible to the caller (paginated, newest first)",
        "description": "Keyset pagination: pass `limit` (1-200, default 50) and the `cursor` returned as `next_cursor` on the previous page. Ordered by created_at DESC with short_code ASC tie-break.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Items per page (1-200, default 50)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque keyset cursor from the previous response's next_cursor",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of URLs plus total count and optional next_cursor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/URLListResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid limit or cursor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/urls/{code}": {
      "parameters": [
        {
          "name": "code",
          "in": "path",
          "required": true,
          "description": "Short code",
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "tags": ["URLs"],
        "summary": "Get URL information",
        "responses": {
          "200": {
            "description": "URL metadata including visit count",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/URLData" }
              }
            }
          },
          "400": {
            "description": "Missing short code",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "URL not found, expired, or owned by someone else",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["URLs"],
        "summary": "Delete a short URL (owner or admin)",
        "responses": {
          "200": {
            "description": "URL deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/URLDeleteResponse" }
              }
            }
          },
          "400": {
            "description": "Missing short code",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "URL not found or expired",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats": {
      "get": {
        "tags": ["URLs"],
        "summary": "Get statistics for visible URLs",
        "responses": {
          "200": {
            "description": "Total URLs and visits",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatsResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all users (paginated, sorted by username)",
        "description": "Admin only. Keyset pagination identical to the URL list (limit 1-200 default 50, cursor via next_cursor).",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Items per page (1-200, default 50)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque keyset cursor from the previous response's next_cursor",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of users plus total count and optional next_cursor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserListResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid limit or cursor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/users/{username}/role": {
      "parameters": [
        {
          "name": "username",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "put": {
        "tags": ["Admin"],
        "summary": "Update a user's role (admin only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["role"],
                "properties": {
                  "role": { "type": "string", "enum": ["admin", "user"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserRoleUpdateResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid body or role",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "The last admin cannot be demoted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/users/{username}": {
      "parameters": [
        {
          "name": "username",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete a user (admin only)",
        "responses": {
          "200": {
            "description": "User deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDeleteResponse" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "The last admin cannot be deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/static/{filepath}": {
      "parameters": [
        {
          "name": "filepath",
          "in": "path",
          "required": true,
          "description": "Relative path of the embedded asset (e.g. app.js), served from /static",
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "tags": ["Operational"],
        "summary": "Static asset for the web UI",
        "description": "Serves the assets embedded at build time (JS, CSS, images) used by the single-page app. Gin registers HEAD for this route as a GET-shaped operation.",
        "security": [],
        "responses": {
          "200": {
            "description": "The asset contents"
          },
          "404": {
            "description": "Asset not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token from /api/v1/auth/login or /api/v1/auth/register. Format: base64url(payload).base64url(signature), HMAC-SHA256 signed."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["message", "code"],
        "properties": {
          "message": { "type": "string", "description": "Human-readable error message" },
          "code": { "type": "string", "description": "Machine-readable error code" }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "password", "email"],
        "properties": {
          "username": { "type": "string", "minLength": 3, "maxLength": 50 },
          "password": { "type": "string", "minLength": 12, "maxLength": 72, "description": "Password length in bytes, 12-72" },
          "email": { "type": "string", "format": "email" }
        }
      },
...
    }
  }
}